pallet_child_bounties::migration

Module v1

Structs

MigrateToV1Impl

Creates a new ids for the child balances based on the child bounty count per parent bounty instead of the total child bounty count. Translates the existing child bounties to the new ids. Creates the V0ToV1ChildBountyIds map from old_child_id to new (parent_id, new_child_id).