pallet_child_bounties

Module pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

The main components of this pallet are:

Pallet, which implements all of the dispatchable extrinsics of the pallet, among other public functions.

The subset of the functions that are dispatchable can be identified either in the dispatchables module or in the Call enum.

storage_types, which contains the list of all types that are representing a storage item. Otherwise, all storage items are listed among Type Definitions.

Config, which contains the configuration trait of this pallet.

Event and Error, which are listed among the Enums.

Modules

dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Structs

Pallet

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

Enums

Call

Contains a variant per dispatchable extrinsic that this pallet has.

Error

The Error enum of this pallet.

Event

The Event enum of this pallet

Traits

Config

Configuration trait of this pallet.

Type Aliases

ChildBounties

Child bounties that have been added.

ChildBountyCount

DEPRECATED: Replaced with ParentTotalChildBounties storage item keeping dedicated counts for each parent bounty. Number of total child bounties. Will be removed in May 2025.

ChildBountyDescriptionsV1

The description of each child-bounty. Indexed by (parent_id, child_id).

ChildrenCuratorFees

The cumulative child-bounty curator fee for each parent bounty.

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

ParentChildBounties

Number of active child bounties per parent bounty. Map of parent bounty index to number of child bounties.

ParentTotalChildBounties

Number of total child bounties per parent bounty, including completed bounties.

V0ToV1ChildBountyIds

The mapping of the child bounty ids from storage version V0 to the new V1 version.