pallet_child_bounties::pallet::dispatchables

Function add_child_bounty

pub fn add_child_bounty<T: Config>(
    parent_bounty_id: BountyIndex,
    value: BalanceOf<T>,
    description: Vec<u8>,
)

Add a new child-bounty.

The dispatch origin for this call must be the curator of parent bounty and the parent bounty must be in “active” state.

Child-bounty gets added successfully & fund gets transferred from parent bounty to child-bounty account, if parent bounty has enough funds, else the call fails.

Upper bound to maximum number of active child bounties that can be added are managed via runtime trait config Config::MaxActiveChildBountyCount.

If the call is success, the status of child-bounty is updated to “Added”.

parent_bounty_id: Index of parent bounty for which child-bounty is being added.

value: Value for executing the proposal.

description: Text description for the child-bounty.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::add_child_bounty.