Function pallet_child_bounties::pallet::dispatchables::award_child_bounty

pub fn award_child_bounty<T: Config>(
    parent_bounty_id: BountyIndex,
    child_bounty_id: BountyIndex,
    beneficiary: <<T as Config>::Lookup as StaticLookup>::Source,
)

Award child-bounty to a beneficiary.

The beneficiary will be able to claim the funds after a delay.

The dispatch origin for this call must be the parent curator or curator of this child-bounty.

Parent bounty must be in active state, for this child-bounty call to work.

Child-bounty must be in active state, for processing the call. And state of child-bounty is moved to “PendingPayout” on successful call completion.

parent_bounty_id: Index of parent bounty.

child_bounty_id: Index of child bounty.

beneficiary: Beneficiary account.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::award_child_bounty.