Function pallet_child_bounties::pallet::dispatchables::close_child_bounty

pub fn close_child_bounty<T: Config>(
    parent_bounty_id: BountyIndex,
    child_bounty_id: BountyIndex,
)

Cancel a proposed or active child-bounty. Child-bounty account funds are transferred to parent bounty account. The child-bounty curator deposit may be unreserved if possible.

The dispatch origin for this call must be either parent curator or T::RejectOrigin.

If the state of child-bounty is Active, curator deposit is unreserved.

If the state of child-bounty is PendingPayout, call fails & returns PendingPayout error.

For the origin other than T::RejectOrigin, parent bounty must be in active state, for this child-bounty call to work. For origin T::RejectOrigin execution is forced.

Instance of child-bounty is removed from the state on successful call completion.

parent_bounty_id: Index of parent bounty.

child_bounty_id: Index of child bounty.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::close_child_bounty.