Module pallet_child_bounties::pallet::storage_types

Auto-generated docs-only module listing all defined storage types for this pallet. Note that members of this module cannot be used directly and are only provided for documentation purposes.

Structs

ChildBounties

Child bounties that have been added.

ChildBountyCount

Number of total child bounties.

ChildBountyDescriptions

The description of each child-bounty.

ChildrenCuratorFees

The cumulative child-bounty curator fee for each parent bounty.

ParentChildBounties

Number of child bounties per parent bounty. Map of parent bounty index to number of child bounties.