pallet_child_bounties

Type Alias BountiesError

pub type BountiesError<T> = Error<T>;

Aliased Type

enum BountiesError<T> {
    InsufficientProposersBalance,
    InvalidIndex,
    ReasonTooBig,
    UnexpectedStatus,
    RequireCurator,
    InvalidValue,
    InvalidFee,
    PendingPayout,
    Premature,
    HasActiveChildBounty,
    TooManyQueued,
    NotProposer,
}

Variants

InsufficientProposersBalance

Proposer’s balance is too low.

InvalidIndex

No proposal or bounty at that index.

ReasonTooBig

The reason given is just too big.

UnexpectedStatus

The bounty status is unexpected.

RequireCurator

Require bounty curator.

InvalidValue

Invalid bounty value.

InvalidFee

Invalid bounty fee.

PendingPayout

A bounty payout is pending. To cancel the bounty, you must unassign and slash the curator.

Premature

The bounties cannot be claimed/closed because it’s still in the countdown period.

HasActiveChildBounty

The bounty cannot be closed because it has active child bounties.

TooManyQueued

Too many approvals are already queued.

NotProposer

User is not the proposer of the bounty.